pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketcanvas-env"
version = "0.1.0"
description = "A minimalist 2D design canvas RL environment with MCP server interface"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "mypy", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.ruff]
target-version = "py310"
line-length = 100
